refactor: migrate to transaction module
