+++ src/actions.ts
@@ -1,3 +1,2 @@
-export * from './actions/helper'
 export * from './actions/domain'
 export * from './actions/record'
